+++ lib/additionals.rb
@@ -55,7 +55,7 @@ module Additionals
       require_dependency 'additionals/hooks'
 
       # Macros
-      load_macros(%w[calendar cryptocompare date fa gist gmap group_users iframe
+      load_macros(%w[calendar cryptocompare date fa gist gmap google_docs group_users iframe
                      issue redmine_issue redmine_wiki
                      last_updated_at last_updated_by meteoblue member new_issue project
                      recently_updated reddit slideshare tradingview twitter user vimeo youtube])