+++ test/functional/wiki_controller_test.rb
@@ -74,6 +74,16 @@ class WikiControllerTest < Additionals::ControllerTest
     assert_select 'iframe[src=?]', '//www.slideshare.net/slideshow/embed_code/57941706'
   end
 
+  def test_show_with_google_docs_macro
+    @request.session[:user_id] = WIKI_MACRO_USER_ID
+    @page.content.text = '{{google_docs(https://docs.google.com/spreadsheets/d/e/RANDOMCODE/pubhtml)}}'
+    @page.content.save!
+    get :show,
+        params: { project_id: 1, id: @page_name }
+    assert_response :success
+    assert_select 'iframe[src=?]', 'https://docs.google.com/spreadsheets/d/e/RANDOMCODE/pubhtml?widget=true&headers=false'
+  end
+
   def test_show_with_iframe_macro
     @request.session[:user_id] = WIKI_MACRO_USER_ID
     @page.content.text = '{{iframe(https://www.redmine.org/)}}'